refactor: Rename CoreHook module example classes from Library to EntryPoint to better reflect their use
